# agents-api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Agents API",
    "description" : "Conversation and messaging service for jorisjonkers.dev",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://agents.jorisjonkers.dev",
    "description" : "Production"
  }, {
    "url" : "http://localhost:8082",
    "description" : "Local development"
  } ],
  "security" : [ {
    "xUserId" : [ ]
  } ],
  "paths" : {
...
    "/api/v1/repositories" : {
      "get" : {
        "tags" : [ "repository-controller" ],
        "operationId" : "list_1",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/RepositoryResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "repository-controller" ],
        "operationId" : "create_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateRepositoryRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RepositoryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repositories/{id}/verify" : {
      "post" : {
        "tags" : [ "repository-controller" ],
        "operationId" : "verify",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RepositoryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repositories/{id}/key" : {
      "post" : {
        "tags" : [ "repository-controller" ],
        "operationId" : "attachKey",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AttachRepositoryDeployKeyRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/v1/projects" : {
      "get" : {
        "tags" : [ "project-controller" ],
        "operationId" : "list_2",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ProjectResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "project-controller" ],
        "operationId" : "create_2",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateProjectRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjectResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/links/{linkId}/key" : {
      "post" : {
        "tags" : [ "project-controller" ],
        "operationId" : "attachKey_1",
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "linkId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AttachDeployKeyRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/v1/projects/{id}/repositories" : {
      "post" : {
        "tags" : [ "project-controller" ],
        "operationId" : "linkRepository",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LinkRepositoryRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/RepositoryResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{id}/links" : {
      "post" : {
        "tags" : [ "project-controller" ],
        "operationId" : "addLink",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AddGithubLinkRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GithubLinkResponse"
                }
              }
            }
          }
        },
        "deprecated" : true
      }
    },
...
    "/api/v1/repositories/{id}" : {
      "get" : {
        "tags" : [ "repository-controller" ],
        "operationId" : "get_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "repository-controller" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/links/{linkId}/setup-guide" : {
      "get" : {
        "tags" : [ "setup-guide-controller" ],
        "operationId" : "guide",
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "linkId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "text/markdown" : {
                "schema" : {
                  "type" : "string"
                }
              },
              "text/plain" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{id}" : {
      "get" : {
        "tags" : [ "project-controller" ],
        "operationId" : "get_2",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{projectId}/links/{linkId}" : {
      "delete" : {
        "tags" : [ "project-controller" ],
        "operationId" : "removeLink",
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "linkId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/v1/projects/{id}/repositories/{repoId}" : {
      "delete" : {
        "tags" : [ "project-controller" ],
        "operationId" : "unlinkRepository",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "repoId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "CreateRepositoryRequest" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "maxLength" : 80,
            "minLength" : 1
          },
          "repoUrl" : {
            "type" : "string",
            "minLength" : 1
          },
          "defaultBranch" : {
            "type" : "string"
          }
        },
        "required" : [ "defaultBranch", "name", "repoUrl" ]
      },
      "AccessVerificationResponse" : {
        "type" : "object",
        "properties" : {
          "read" : {
            "type" : [ "boolean", "null" ]
          },
          "write" : {
            "type" : [ "boolean", "null" ]
          },
          "defaultBranchProtected" : {
            "type" : [ "boolean", "null" ]
          },
          "checkedAt" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          },
          "messages" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        },
        "required" : [ "messages" ]
      },
      "RepositoryResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "repoUrl" : {
            "type" : "string"
          },
          "defaultBranch" : {
            "type" : "string"
          },
          "vaultKeyPath" : {
            "type" : "string"
          },
          "deployKeyFingerprint" : {
            "type" : [ "string", "null" ]
          },
          "deployKeyAddedAt" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "verification" : {
            "oneOf" : [ {
              "$ref" : "#/components/schemas/AccessVerificationResponse"
            }, {
              "type" : "null"
            } ]
          }
        },
        "required" : [ "createdAt", "defaultBranch", "id", "name", "repoUrl", "updatedAt", "vaultKeyPath" ]
      },
      "AttachRepositoryDeployKeyRequest" : {
        "type" : "object",
        "properties" : {
          "privateKeyOpenssh" : {
            "type" : "string",
            "minLength" : 1
          },
          "publicKeyOpenssh" : {
            "type" : "string",
            "minLength" : 1
          },
          "knownHosts" : {
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "privateKeyOpenssh", "publicKeyOpenssh" ]
      },
      "CreateProjectRequest" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "maxLength" : 80,
            "minLength" : 1
          },
          "slug" : {
            "type" : "string",
            "pattern" : "^[a-z0-9][a-z0-9-]{0,62}$"
          },
          "description" : {
            "type" : [ "string", "null" ],
            "maxLength" : 1000,
            "minLength" : 0
          }
        },
        "required" : [ "name", "slug" ]
      },
      "ProjectResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "slug" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "required" : [ "createdAt", "description", "id", "name", "slug", "updatedAt" ]
      },
      "AttachDeployKeyRequest" : {
        "type" : "object",
        "properties" : {
          "privateKeyOpenssh" : {
            "type" : "string",
            "minLength" : 1
          },
          "publicKeyOpenssh" : {
            "type" : "string",
            "minLength" : 1
          },
          "knownHosts" : {
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "privateKeyOpenssh", "publicKeyOpenssh" ]
      },
      "LinkRepositoryRequest" : {
        "type" : "object",
        "properties" : {
          "repositoryId" : {
            "type" : "string",
            "format" : "uuid"
          }
        },
        "required" : [ "repositoryId" ]
      },
      "AddGithubLinkRequest" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "minLength" : 1
          },
          "repoUrl" : {
            "type" : "string",
            "minLength" : 1
          },
          "defaultBranch" : {
            "type" : "string"
          }
        },
        "required" : [ "defaultBranch", "name", "repoUrl" ]
      },
      "GithubLinkResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "projectId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "repoUrl" : {
            "type" : "string"
          },
          "defaultBranch" : {
            "type" : "string"
          },
          "vaultKeyPath" : {
            "type" : "string"
          },
          "deployKeyFingerprint" : {
            "type" : [ "string", "null" ]
          },
          "deployKeyAddedAt" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "required" : [ "createdAt", "defaultBranch", "id", "name", "projectId", "repoUrl", "updatedAt", "vaultKeyPath" ]
      },
...
    },
    "securitySchemes" : {
      "xUserId" : {
        "type" : "apiKey",
        "description" : "User identifier injected by Traefik forward-auth",
        "name" : "X-User-Id",
        "in" : "header"
      }
    }
  }
}